pyproject.toml
[tool.poetry]
name = "orange-bank-api"
version = "0.1.0"
description = ""
authors = ["Alec McGavin <alec@mcgav.in>"]

[tool.poetry.dependencies]
python = "^3.7"
pycryptodomex = "^3.9.4"
requests = "^2.22.0"
pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
ipython = "^7.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
